+++ test/macroy/org_file_test.exs
@@ -1,0 +1,39 @@
+defmodule Macroy.OrgFileTest do
+  use Macroy.DataCase
+  alias Macroy.{OrgFile, Todo}
+
+  describe "read/1" do
+    test "valid orgfile returns valid todos" do
+      orgfile = %OrgFile{
+        host: "localhost",
+        filename: "usda_prime_org_file.org",
+        path: "/"
+      }
+      orgfiles = OrgFile.read(orgfile)
+      with todo_1 <- Enum.at(orgfiles, 0),
+           todo_2 <- Enum.at(orgfiles, 1) do
+        # Todo 1 and 2 are Todos{} too
+        assert %Todo{} = todo_1
+        assert %Todo{} = todo_2
+
+        # Todo 1
+        assert todo_1.name == "Go over Google Assistant stuff in email"
+        assert todo_1.category == "Electronic Organization"
+        assert is_nil(todo_1.subcategory)
+        assert is_nil(todo_1.closed_on)
+        assert is_nil(todo_1.deadline_on)
+        assert DateTime.compare(todo_1.scheduled_for, ~U[2019-08-06 04:00:00Z]) == :eq
+        refute todo_1.is_done
+
+        # Todo 2
+        assert todo_2.name == "Go play in an Elixir stream"
+        assert todo_2.category == "Education"
+        assert todo_2.subcategory == "Coding"
+        assert DateTime.compare(todo_2.closed_on,  ~U[2019-08-18 14:50:00Z]) == :eq
+        assert DateTime.compare(todo_2.deadline_on, ~U[2019-08-13 04:00:00Z]) == :eq
+        assert DateTime.compare(todo_2.scheduled_for, ~U[2019-08-13 04:00:00Z]) == :eq
+        assert todo_2.is_done
+      end
+    end
+  end
+end